Add wish service and model. Add some methods to work with data from DB
