+++ src/main/java/com/epam/orderingsystem/service/WishBuilderService.java
@@ -1,0 +1,11 @@
+package com.epam.orderingsystem.service;
+
+import com.epam.orderingsystem.model.Child;
+import com.epam.orderingsystem.model.GiftOrder;
+import com.epam.orderingsystem.model.Wish;
+
+import java.util.List;
+
+public interface WishBuilderService {
+    List<Wish> build(List<Child> children, List<GiftOrder> orders);
+}
